Refactoring to   to improve readability , modularity and encapsulation of the components of Feature 2 : Outdoor Directions.  The component is working as intended after manual test.
__Substitute Algorithm__
Problem : The logic was added to the outdoor screen , which is only meant to be used as view , the logic has been removed and set into it's own component

Created directionLogic.tsx and placed it in maps folder

__Encapsulate Variable__

Created the DirectionFields.tsx and placed it in /ui folder
__

Now:
The screen imports the ui component for the origin/direction fields
The screen uses the hook useDirectionLogic for the logic.
I updated the snapshot for outdoorDirection screen since the tree would not match anymore

Problem : snapshot is failing to pass with error "Warning: Function components cannot be given refs. Attempts to access this ref will fail. Did you mean to use React.forwardRef()?"

I will fix this error in another commit.