clicking on a building polygon  allows you to set it as destination OutdoorDirection Screen is a composite of DirectionFields and OutdoorMap OutdoorMap is a composite containing polygonRender PolygonRender accesses building_outlines to get the coordinates.
Functionalities were added in the useDirectionLogic
- setToBuildingLocation  :  will set the start or destination field to the buildingLocation
- setToBuildingLocation is called if buildingLocation changes

In polygonRender , onPress the function buildingAsDestination is called.
buildingAsDestination prompts the user to know if they want to make this building the destination.
if result is returned true by buildingAsDestination then polygonRender uses setBuildingDestination as the center of the polygon.   (refactoring later : it would be better if it lead to the entrance but the entrance were not included in the building_outlines file ).